Added h3 tag
This is to add a title to the very top of the page, above the toolbar, as the branch suggests. To keep the current format, I just moved the properties of the previous h1 to h2, and the previous properties of h2 to h3.

I also fixed some spacing issues around the rest of the file.